Remote filterText padding
This was never exceptionall consistent across browsers, but it seems that recent typography changes have made that difference more noticeable. Since we're not using a "normal" dropdown opener for the filter menu, it makes the most sense to try to match it to the "normal" one (i.e. the sort menu). The best way to do this is to not add any padding - this does mean there's not a lot of space between the chevron on text, but at least the gap is the same between the two menus for any given browser.