Decouple d2l-all-courses
Rather than passing in the pinnedEnrollments, unpinnedEnrollments, and myEnrollmentsEntity objects as parameters to d2l-all-courses, instead just pass in the enrollmentsSearchUrl. This does mean we're probably re-doing the same fetch as when we first loaded the widget, but a) caching, and b) we no longer need a 600ms delay to avoid animation stutter, so this is actually faster (assuming the call takes <600ms, which is very usually does). Another benefit of this approach is that d2l-my-courses itself can now be completely ignorant of unpinned enrollments (future PR though, to keep things cleaner).