Decouple d2l-all-courses and d2l-my-courses
Rather than directly setting the pinned and unpinned courses, we can instead treat the connection between d2l-all-courses and d2l-my-courses like the connection between d2l-all-courses and the filter - just set the myEnrollmentsEntity, and let the child component take it from there. The downside to this is that we do need to re-fetch things, but a) caching, and b) we no longer require the 600ms delay to make the animation not stutter.