Set _searchUrl in load with bustCache
This isn't strictly related, but it does fix a bug that becomes more obvious with this approach. If you pin/unpin courses in the overlay, then close and reopen it, the changes you made don't appear. This is because d2l-fetch returns the same entity from the initial fetch, which has the same URL. We don't have a way to invalidate the cache directly, so instead use the ?bustCache method used elsewhere.